refactor: move actions-lib to packages/actions-core
- Move .github/actions/lib to packages/actions-core (whole directory)
- Rename @code-assistants/actions-lib to @code-assistants/actions-core
- Replace .github/actions/lib entry in workspaces with packages/* glob
- Update import paths in agents-rules-sync.ts and changeDetector.ts
- Regenerate bun.lock after workspace rename
