+++ package.json
@@ -16,10 +16,10 @@
     "language": "typescript"
   },
   "workspaces": [
-    ".github/actions/lib",
     ".github/actions/files-sync",
     ".github/actions/agents-rules-sync",
-    ".github/actions/release-action"
+    ".github/actions/release-action",
+    "packages/*"
   ],
   "scripts": {
     "prepare": "husky",
